Now Playing scripts
np.sh pulls the RSS feed for last.fm user CoCoMpls, finesses it a bit,
and saves it to a text file. nowplayingv1.py is a slight variation on
txt2litebrite.py, specifically taking the bottom 3 lines of the board,
prefixing "Now Playing", "Last Played", and "Double Prev", alternating
colors, and expiring its lease as soon as it's done.

Plenty of room for improvements:
1. Do everything in python
2. Use peggy V2 (when it's stable)
3. Add logic to avoid unnecessary refreshes
4. Add logic to avoid excessive last.fm polling